+++ tybotRoute/models/TiledeskChatbotUtil.js
@@ -518,20 +518,39 @@ class TiledeskChatbotUtil {
           }
         }
         
-
+        const _bot = chatbot.bot; // aka FaqKB
+        if (chatbot.log) {
+            console.log("Adding Globals to context..., chatbot.attributes?", JSON.stringify(_bot));
+        }
+        
+        if (_bot.attributes && _bot.attributes.globals) {
+            if (chatbot.log) {console.log("Got Globals:", JSON.stringify(_bot.attributes.globals));}
+            _bot.attributes.globals.forEach(async (global_var) => {
+                if (chatbot.log) {console.log("Adding global:", global_var.key, "value:", global_var.value);}
+                await chatbot.addParameter(global_var.key, global_var.value);
+            });
+        }
+        // await chatbot.addParameter("testVar",
+        //     {
+        //         name: "Andrea",
+        //         coords: {
+        //             x: 2, y: 1
+        //         }
+        //     }
+        // );
         
         if (chatbot.log) {
           // console.log("tdcache:", chatbot.tdcache);
           console.log("requestId:", requestId);
           console.log("KEY:", TiledeskChatbotConst.REQ_PROJECT_ID_KEY);
-          console.log("TiledeskChatbot:", TiledeskChatbot);
+        //   console.log("TiledeskChatbot:", TiledeskChatbot);
           let proj_ = await chatbot.getParameter(TiledeskChatbotConst.REQ_PROJECT_ID_KEY);
           console.log("request parameter proj_:", proj_);
           const all_parameters = await chatbot.allParameters();
           for (const [key, value] of Object.entries(all_parameters)) {
             // const value = all_parameters[key];
             const value_type = typeof value;
-            if (chatbot.log) {console.log("request parameter:", key, "value:", value, "type:", value_type)}
+            if (chatbot.log) {console.log("REQUEST ATTRIBUTE:", key, "VALUE:", value, "TYPE:", value_type)}
           }
         }
     } catch(error) {